Add Latin subsets to Poppins font and adjust CSS margins
Added "latin" to the subsets of the Poppins font in the _app.tsx file for broader language support. Also, a bottom margin was added to the hero module text style for improved layout and readability. Furthermore, TaskOptions in watcherTasks.xml is now disabled by setting isEnabled to false for optimized development environment.